migrating from Junit to TestNG (RealItemTests, VirtualItemTests classes) adding 3 xml files
