+++ src/test/java/task20/all.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="smoke suite">
+    <test name="shop-test">
+        <packages>
+            <package name="task20" />
+        </packages>
+    </test>
+</suite>
