+++ src/test/java/task20/regression.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="regression suite">
+    <test name="shop-test">
+        <groups>
+            <run>
+                <include name="regression"/>
+            </run>
+        </groups>
+        <classes>
+            <class name = "task20.CartTests"/>
+            <class name = "task20.ParserTests"/>
+            <class name = "task20.RealItemTests"/>
+            <class name = "task20.VirtualItemTests"/>
+        </classes>
+    </test>
+</suite>
