--- src/test/java/testng.xml
@@ -1,12 +1,0 @@
-<?xml version = "1.0" encoding = "UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name = "Suite1">
-    <test name = "Test1">
-        <classes>
-          <!--  <class name = "CartTests" />-->
-            <class name = "ParserTests" />
-        </classes>
-
-    </test>
-</suite>
